Removed indirect dependency between the interceptor and ObjectProperty class
