+++ src/main/java/com/github/sylvainlaurent/maven/swaggervalidator/doc/traversal/node/ObjectPropertyWrapper.java
@@ -1,9 +1,7 @@
 package com.github.sylvainlaurent.maven.swaggervalidator.doc.traversal.node;
 
-import static java.util.Collections.synchronizedMap;
 import static org.apache.commons.collections4.ListUtils.emptyIfNull;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -14,7 +12,6 @@ import io.swagger.models.properties.Property;
 
 public class ObjectPropertyWrapper implements VisitableProperty {
 
-    private static final Map<String, List<String>> requiredObjectPropertiesMap = synchronizedMap(new HashMap<String, List<String>>());
 
     private final ObjectProperty objectProperty;
     private final String propertyName;
@@ -35,19 +32,14 @@ public class ObjectPropertyWrapper implements VisitableProperty {
     }
 
     public List<String> getRequiredProperties() {
-        return emptyIfNull(requiredObjectPropertiesMap.get(key(objectProperty)));
+        return emptyIfNull(ObjectPropertyRequiredStore.get(objectProperty));
     }
 
     public Map<String, Property> getProperties() {
         return objectProperty.getProperties();
     }
 
-    public static void storeRequiredProperties(Object objectProperties, List<String> required) {
-        requiredObjectPropertiesMap.put(key(objectProperties), required);
-    }
 
-    private static String key(Object objectProperties) {
-        return String.valueOf(System.identityHashCode(objectProperties));
-    }
+
 
 }
